+++ src/react/ChatInputActionBar/demos/index.tsx
@@ -20,9 +20,9 @@ export default () => {
             {
               icon: GlobeIcon,
               key: 'search',
-              wrapper: (node, key) => {
+              wrapper: (node) => {
                 return (
-                  <Popover arrow={false} content={'Test Popover'} key={key}>
+                  <Popover arrow={false} content={'Test Popover'}>
                     {node}
                   </Popover>
                 );