+++ src/react/ChatInputActions/demos/Collapse.tsx
@@ -14,9 +14,9 @@ const items: ChatInputActionsProps['items'] = [
   {
     icon: GlobeIcon,
     key: 'search',
-    wrapper: (node, key) => {
+    wrapper: (node) => {
       return (
-        <Popover arrow={false} content={'Test Popover'} key={key}>
+        <Popover arrow={false} content={'Test Popover'}>
           {node}
         </Popover>
       );