+++ src/react/ChatInputActions/type.ts
@@ -4,12 +4,12 @@ import type { FlexboxProps } from 'react-layout-kit';
 
 export type ChatInputActionEvent = Pick<MenuInfo, 'key' | 'keyPath' | 'domEvent'>;
 
-export type ActionItem = MenuItemType & {
+export interface ActionItem extends MenuItemType {
   active?: boolean;
   alwaysDisplay?: boolean;
   children?: ReactNode;
-  wrapper?: (dom: ReactNode, key: string) => ReactNode;
-};
+  wrapper?: (dom: ReactNode) => ReactNode;
+}
 
 export type DividerItem = {
   type: 'divider';
@@ -17,9 +17,6 @@ export type DividerItem = {
 
 export type CollapseItem = {
   children: (ActionItem | DividerItem)[];
-  defaultExpand?: boolean;
-  expand?: boolean;
-  onChange?: (expand: boolean) => void;
   type: 'collapse';
 };
 
@@ -31,17 +28,21 @@ export type DropdownItem = Omit<ActionItem, 'children' | 'type'> & {
 export type ChatInputActionItem = ActionItem | DividerItem | CollapseItem | DropdownItem;
 
 export interface ChatInputActionsProps extends Omit<FlexboxProps, 'children'> {
+  autoCollapse?: boolean;
   collapseOffset?: number;
+  defaultGroupCollapse?: boolean;
   disabled?: boolean;
+  groupCollapse?: boolean;
   items?: ChatInputActionItem[];
   onActionClick?: (action: ChatInputActionEvent) => void;
+  onGroupCollapseChange?: (collapse: boolean) => void;
 }
 
 export interface ChatInputActionsCollapseProps {
   children?: ReactNode;
-  defaultExpand?: boolean;
-  expand?: boolean;
+  collapse?: boolean;
   gap?: FlexboxProps['gap'];
+  groupCollapse?: boolean;
   mode?: 'default' | 'popup';
-  onChange?: (expand: boolean) => void;
+  onGroupCollapseChange?: (collapse: boolean) => void;
 }