+++ CHANGELOG.md
@@ -2,6 +2,99 @@
 
 # Changelog
 
+### [Version&nbsp;1.4.7](https://github.com/lobehub/lobe-editor/compare/v1.4.6...v1.4.7)
+<sup>Released on **2025-09-07**</sup>
+
+
+#### 🐛 Bug Fixes
+
+- **misc**: Fix ChatInputActions autoCollapse.
+
+
+<br/>
+
+
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+
+
+#### What's fixed
+
+* **misc**: Fix ChatInputActions autoCollapse ([249869e](https://github.com/lobehub/lobe-editor/commit/249869e))
+
+</details>
+
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+### [Version&nbsp;1.4.6](https://github.com/lobehub/lobe-editor/compare/v1.4.5...v1.4.6)
+<sup>Released on **2025-09-06**</sup>
+
+
+#### 🐛 Bug Fixes
+
+- **misc**: Remove auto Collapsed temporarily to fix bounce.
+
+
+<br/>
+
+
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+
+
+#### What's fixed
+
+* **misc**: Remove auto Collapsed temporarily to fix bounce, closes [#18](https://github.com/lobehub/lobe-editor/issues/18) ([4118154](https://github.com/lobehub/lobe-editor/commit/4118154))
+
+</details>
+
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
+### [Version&nbsp;1.4.5](https://github.com/lobehub/lobe-editor/compare/v1.4.4...v1.4.5)
+<sup>Released on **2025-09-05**</sup>
+
+
+#### 💄 Styles
+
+- **misc**: Update hr magrin.
+
+
+<br/>
+
+
+
+<details>
+<summary><kbd>Improvements and Fixes</kbd></summary>
+
+
+
+#### Styles
+
+* **misc**: Update hr magrin ([ff932c1](https://github.com/lobehub/lobe-editor/commit/ff932c1))
+
+</details>
+
+
+<div align="right">
+
+[![](https://img.shields.io/badge/-BACK_TO_TOP-151515?style=flat-square)](#readme-top)
+
+</div>
+
 ### [Version&nbsp;1.4.4](https://github.com/lobehub/lobe-editor/compare/v1.4.3...v1.4.4)
 <sup>Released on **2025-09-05**</sup>